requirements.txt
streamlit
numpy
faiss-cpu
openai
PyYAML
sentence-transformers
llama-cpp-python
